Refactor: 코드 구현 내용 패키지 변경(user->book)
