+++ src/main/java/teamssavice/ssavice/user/controller/UserController.java
@@ -50,14 +50,4 @@ public class UserController {
 
         return ResponseEntity.ok(UserResponse.Summary.from(model));
     }
-
-    @GetMapping("/book/summary")
-    public ResponseEntity<UserResponse.BookSummary> getBookSummary(
-        @CurrentId Long userId
-    ) {
-        UserModel.BookSummary model = userService.getBookSummary(userId);
-
-        return ResponseEntity.ok(UserResponse.BookSummary.from(model));
-    }
-
 }
